ml_services/infra_redis_zmq/requirements.txt
pyzmq==25.1.1
redis==4.6.0
python-dateutil==2.8.2
